pyproject.toml
[tool.poetry]
name = "git-whisper"
version = "0.1.0"
description = "Local commit message generator using ollama"
authors = ["haradama <harada@sfc.wide.ad.jp>"]
homepage = "https://github.com/haradama"
repository = "https://github.com/haradama/git-whisper"
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
gitpython = "^3.1.44"
ollama = "^0.4.7"
pydantic = "^2.10.6"

[tool.poetry.scripts]
git-whisper = "git_whisper.cli:main"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-mock = "^3.14.0"
pytest-cov = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
